Protect against address arithmetic overflow.

+++ gettext-runtime/intl/wprintf-parse.h
@@ -21,6 +21,9 @@
 
 #include "printf-args.h"
 
+/* Get ssize_t.  */
+#include <sys/types.h>
+
 
 /* Flags */
 #define FLAG_GROUP	 1	/* ' flag */
@@ -38,22 +41,22 @@ typedef struct
   int flags;
   const wchar_t* width_start;
   const wchar_t* width_end;
-  int width_arg_index;
+  ssize_t width_arg_index;
   const wchar_t* precision_start;
   const wchar_t* precision_end;
-  int precision_arg_index;
+  ssize_t precision_arg_index;
   wchar_t conversion; /* d i o u x X f e E g G c s p n U % but not C S */
-  int arg_index;
+  ssize_t arg_index;
 }
 wchar_t_directive;
 
 /* A parsed format string.  */
 typedef struct
 {
-  unsigned int count;
+  size_t count;
   wchar_t_directive *dir;
-  unsigned int max_width_length;
-  unsigned int max_precision_length;
+  size_t max_width_length;
+  size_t max_precision_length;
 }
 wchar_t_directives;
